Final changes

+++ Proyecto final/Simple Tasks/Simple Tasks/DBManagement.h
@@ -38,7 +38,7 @@
 - (void) modifyTask: (id) currentTask newTask: (id) modifiedTask taskListTitle: (NSString *) title;
 - (void) deleteTask:(id) taskToBeDeleted parentTaskList: (NSString *) taskListTitle;
 - (void) setTaskCompleted: (id) currentTask taskDone: (NSNumber *) taskCompletion parentTaskList: (NSString *) pTaskList;
-- (int) countTasks: (NSString *) taskListTitle;
+- (NSUInteger) countTasks: (NSString *) taskListTitle;
 
 // Task List
 - (void) addTaskList: (id) taskListInfo tasks: (NSArray *) myTasks;
